PokerSimulator
Hi, this project is designed mainly for the prototype of a poker simulator, first of all, let me introduce the rules of this project.

You, and your friend are given 27 pokers, respectively. And one of you have to place the top of your pokers on the table and then the other one as well,(in vertical) if the number of poker you`re going to place is correspond with any poker on the table, you can have the first poker which has the same number with the poker you`re going to place and all of the pokers below the first one. once a person lose all of the pokers he has, he lose the game.

This project is only a prototype, therefore no output will display on the winform, if you wish to see the output, please set the breakpoint and fellow the process.